+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 Fixes:
 
-- Type annotation fixes for the Firestore connector
+- Type annotation fixes in workflowslib
 
 ## Version 0.13.0 - 2025-10-12